+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Changed
 
+  - Deployment modal - Add credential check
   - Deployment modal - Filter infrastructure services by module subtype
-  - Deployment modal - Add connectivity check popup
   - Deployment modal - Credential step is replaced by infrastructure services step
   - Session page should not appear when reloading the page
   - DASHBOARD DETAIL PAGE - ACL widget disapearing sometimes on refresh fix